# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OrionOps API
  description: Enterprise Service Orchestration Platform - ITSM + ERP Extensions
  version: 1.0.0
  contact:
    name: OrionOps Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /workflows/instances/{id}/tasks:
    get:
      tags: [Workflow]
      summary: Get workflow tasks
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Task list
  /workflows/tasks/{taskId}/complete:
    post:
      tags: [Workflow]
      summary: Complete workflow task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task completed

  # ---- Finance ----
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    EntityIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Entity UUID
...